% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jitter_power.R
\name{jitter_power}
\alias{jitter_power}
\title{Function to Estimate the Power of a Spatial Relative Risk using previously
collected data.}
\usage{
jitter_power(
  obs_data,
  sim_total,
  samp_control = c("uniform", "CSR", "MVN"),
  s_control = 1,
  upper_tail = 0.975,
  lower_tail = 0.025,
  cascon = FALSE,
  resolution = 128,
  edge = "uniform",
  adapt = FALSE,
  h0 = NULL,
  verbose = TRUE,
  parallel = FALSE,
  n_core = NULL,
  ...
)
}
\arguments{
\item{obs_data}{}

\item{sim_total}{}

\item{samp_control}{}

\item{s_control}{}

\item{upper_tail}{}

\item{lower_tail}{}

\item{cascon}{}

\item{resolution}{}

\item{edge}{}

\item{adapt}{}

\item{h0}{}

\item{verbose}{}

\item{parallel}{}

\item{n_core}{}

\item{...}{}
}
\value{

}
\description{
Function to Estimate the Power of a Spatial Relative Risk using previously
collected data.
}
