% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_data.R
\name{spatial_data}
\alias{spatial_data}
\title{Function to Simulate and Combine User-Specified Case Clusters and Randomized Control Clusters}
\usage{
spatial_data(
  x_case,
  y_case,
  x_control = NULL,
  y_control = NULL,
  n_case = NULL,
  n_control = NULL,
  npc_control = NULL,
  r_case = NULL,
  r_control = NULL,
  s_case = NULL,
  s_control = NULL,
  l_case = NULL,
  l_control = NULL,
  e_control = NULL,
  sim_total,
  samp_case = c("uniform", "MVN", "CSR", "IPP"),
  samp_control = c("uniform", "systematic", "MVN", "CSR", "IPP", "clustered"),
  win = unit.square(),
  ...
)
}
\arguments{
\item{x_case}{TO ADD}

\item{y_case}{TO ADD}

\item{x_control}{TO ADD}

\item{y_control}{TO ADD}

\item{n_case}{TO ADD}

\item{n_control}{TO ADD}

\item{npc_control}{TO ADD}

\item{r_case}{TO ADD}

\item{r_control}{TO ADD}

\item{s_case}{TO ADD}

\item{s_control}{TO ADD}

\item{l_case}{TO ADD}

\item{l_control}{TO ADD}

\item{e_control}{TO ADD}

\item{sim_total}{TO ADD}

\item{samp_case}{TO ADD}

\item{samp_control}{TO ADD}

\item{win}{TO ADD}

\item{...}{TO ADD}
}
\value{
TO ADD, add output of the function here
}
\description{
Function to Simulate and Combine User-Specified Case Clusters and Randomized Control Clusters
}
\examples{
\dontrun{
 spatial_data(x_case = c(0.25, 0.5, 0.75),
              y_case = c(0.75, 0.25, 0.75),
              x_control = c(0.25, 0.5, 0.75),
              y_control = c(0.75, 0.25, 0.75),
                 # n_case = c(100, 100,100),
              n_case = 100,
              n_control = 700,
                 # r_case = c(0.1, 0.2, 0.1),
              r_case = 0.1,
                 # s_case = c(0.05,0.01,0.05),
              s_case = 0.05,
                 # l_case = c(200,100,200),
              l_case = l_cont,
              sim_total = 2,
              samp_case = "MVN", 
              samp_control = "MVN",
              npc_control = 10,
              r_control = 0.1,
              e_control = 0,
                 # l_control = 100,
              l_control = l_cont,
              s_control = 0.1,
              win = unit.circle
              )
}

}
