% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_plots.R
\name{spatial_plots}
\alias{spatial_plots}
\title{Function to Estimate the Power of a Spatial Relative Risk using Simulated Data}
\usage{
spatial_plots(
  input,
  p_thresh = 0.8,
  plot_text = FALSE,
  n_sim = 4,
  cols = c("#000000", "#cccccc", "ffffff", "ff0000", "0000ff"),
  chars = c(1, 1),
  sizes = c(0.1, 0.1),
  plot_pts = TRUE,
  ...
)
}
\arguments{
\item{input}{}

\item{p_thresh}{}

\item{plot_text}{}

\item{n_sim}{}

\item{cols}{}

\item{chars}{}

\item{sizes}{}

\item{plot_pts}{}

\item{...}{}
}
\value{

}
\description{
Function to Estimate the Power of a Spatial Relative Risk using Simulated Data
}
