% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_power.R
\name{spatial_power}
\alias{spatial_power}
\title{Function to Estimate the Power of a Spatial Relative Risk using Simulated Data}
\usage{
spatial_power(
  x_case,
  y_case,
  x_control = NULL,
  y_control = NULL,
  n_case = NULL,
  n_control = NULL,
  npc_control = NULL,
  r_case = NULL,
  r_control = NULL,
  s_case = NULL,
  s_control = NULL,
  l_case = NULL,
  l_control = NULL,
  e_control = NULL,
  sim_total = 1,
  samp_case = c("uniform", "MVN", "CSR", "IPP"),
  samp_control = c("uniform", "systematic", "MVN", "CSR", "IPP", "clustered"),
  upper_tail = 0.975,
  lower_tail = 0.025,
  win = unit.square(),
  cascon = FALSE,
  resolution = 128,
  edge = "uniform",
  adapt = FALSE,
  h0 = NULL,
  verbose = TRUE,
  parallel = FALSE,
  n_core = NULL,
  ...
)
}
\arguments{
\item{x_case}{}

\item{y_case}{}

\item{x_control}{}

\item{y_control}{}

\item{n_case}{}

\item{n_control}{}

\item{npc_control}{}

\item{r_case}{}

\item{r_control}{}

\item{s_case}{}

\item{s_control}{}

\item{l_case}{}

\item{l_control}{}

\item{e_control}{}

\item{sim_total}{}

\item{samp_case}{}

\item{samp_control}{}

\item{upper_tail}{}

\item{lower_tail}{}

\item{win}{}

\item{cascon}{}

\item{resolution}{}

\item{edge}{}

\item{adapt}{}

\item{h0}{}

\item{verbose}{}

\item{parallel}{}

\item{n_core}{}

\item{...}{}
}
\value{

}
\description{
Function to Estimate the Power of a Spatial Relative Risk using Simulated Data
}
