% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jitter_power.R
\name{jitter_power}
\alias{jitter_power}
\title{jitter_power()}
\usage{
jitter_power(
  obs_data,
  sim_total,
  samp_control = c("uniform", "CSR", "MVN"),
  s_control = 1,
  upper_tail = 0.975,
  lower_tail = 0.025,
  cascon = FALSE,
  resolution = 128,
  edge = "uniform",
  adapt = FALSE,
  h0 = NULL,
  verbose = TRUE,
  parallel = FALSE,
  n_core = NULL,
  ...
)
}
\arguments{
\item{obs_data}{TO ADD}

\item{sim_total}{Number of simulation iterations}

\item{samp_control}{Type of random sampling for controls ('CSR', 'uniform', 'MVN')}

\item{s_control}{If MVN, the standard deviation of the random normal noise added to each coordinate of the control locations}

\item{upper_tail}{User-specified upper tail of a two-tailed significance level}

\item{lower_tail}{User-specified lower tail of a two-tailed significance level}

\item{cascon}{TRUE for power to detect both relative case and control clustering (hot and coldspots)}

\item{resolution}{10 to calculate surfaces in a 10 x 10 grid}

\item{edge}{"diggle" to employ the Diggle method that reweights each observation-specific kernel, default is "uniform"}

\item{adapt}{FALSE to estimate using fixed smoothing, future direction: explore adaptive smoothing}

\item{h0}{NULL for internal estimation of a common oversmoothing bandwidth computed via the \code{sparr::OS()} function in the sparr package, can be user specified if want to force same bandwidth across iterations}

\item{verbose}{TO ADD}

\item{parallel}{TO ADD}

\item{n_core}{TO ADD}

\item{...}{TO ADD}
}
\value{
TO ADD, add output of the function here
}
\description{
Function to Estimate the Power of a Spatial Relative Risk using previously
collected data.
}
\details{
\itemize{
\item There are other arguments for tuning (mostly for adaptive smoothing), see
\code{sparr::risk()} helpfile.
\item NOTE: Force the \code{sparr::risk()} arguement tolerate = TRUE to always calculate asymptotic p-vlaue surfaces
\item NOTE: Force the \code{sparr::risk()} arguement verbose = FALSE to clean-up presentation
}
}
\examples{
## From the 'sparr' package
\dontrun{
data(pbc)
sim_power <- jitter_power(obs_data = pbc,
                          sim_total = 100,
                          samp_control = "MVN",
                          parallel = TRUE,
                          s_control = 10, # default = 1
                          upper_tail = 0.995, # default = 0.975
                          lower_tail = 0.005, # default = 0.025
                          resolution = 100, # default = 128
                          edge = "diggle", # default = "uniform"
                          cascon = FALSE # default = FALSE
                         )
}

}
