% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_data.R
\name{spatial_data}
\alias{spatial_data}
\title{Function to Simulate and Combine User-Specified Case Clusters and Randomized Control Clusters}
\usage{
spatial_data(
  x_case,
  y_case,
  x_control = NULL,
  y_control = NULL,
  n_case = NULL,
  n_control = NULL,
  npc_control = NULL,
  r_case = NULL,
  r_control = NULL,
  s_case = NULL,
  s_control = NULL,
  l_case = NULL,
  l_control = NULL,
  e_control = NULL,
  sim_total,
  samp_case = c("uniform", "MVN", "CSR", "IPP"),
  samp_control = c("uniform", "systematic", "MVN", "CSR", "IPP", "clustered"),
  win = unit.square(),
  ...
)
}
\arguments{
\item{x_case}{}

\item{y_case}{}

\item{x_control}{}

\item{y_control}{}

\item{n_case}{}

\item{n_control}{}

\item{npc_control}{}

\item{r_case}{}

\item{r_control}{}

\item{s_case}{}

\item{s_control}{}

\item{l_case}{}

\item{l_control}{}

\item{e_control}{}

\item{sim_total}{}

\item{samp_case}{}

\item{samp_control}{}

\item{win}{}

\item{...}{}
}
\value{

}
\description{
Function to Simulate and Combine User-Specified Case Clusters and Randomized Control Clusters
}
