% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_plots.R
\name{spatial_plots}
\alias{spatial_plots}
\title{Function to Estimate the Power of a Spatial Relative Risk using Simulated Data}
\usage{
spatial_plots(
  input,
  p_thresh = 0.8,
  plot_text = FALSE,
  n_sim = 4,
  cols = c("#000000", "#cccccc", "ffffff", "ff0000", "0000ff"),
  chars = c(1, 1),
  sizes = c(0.1, 0.1),
  plot_pts = TRUE,
  ...
)
}
\arguments{
\item{input}{Use output of SRR simulation}

\item{p_thresh}{p threshold, default is 0.8}

\item{plot_text}{default is FALSE, in case resolution >> 10}

\item{n_sim}{TO ADD}

\item{cols}{colors, default colors = c("grey0", "grey80", "grey100", "red", "blue")}

\item{chars}{case, control}

\item{sizes}{case, control}

\item{plot_pts}{TO ADD, default is TRUE}

\item{...}{TO ADD}
}
\value{
TO ADD, add output of the function here
}
\description{
Function to Estimate the Power of a Spatial Relative Risk using Simulated Data
}
\examples{
\dontrun{
# Default colors = c("grey0", "grey80", "grey100", "red", "blue")
spatial_plots(input = sim_power, # use output of SRR simulation
              p_thresh = 0.8, # default = 0.8
              #plot_text = T, # default = FALSE in case resolution >> 10
              plot_pts = T, # default = TRUE 
              chars = c(4,5), # case, control
              sizes = c(0.5,0.5), # case, control
              cols = c("#0000ff", "#00ff00", "#ff0000", "#a020f0", "#ffa500") #c("blue", "green", "red", "purple", "orange") # insufficient, sufficient, text, case, control
              )
}

}
