% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_plots.R
\name{spatial_plots}
\alias{spatial_plots}
\title{spatial_plots()}
\usage{
spatial_plots(
  input,
  p_thresh = 0.8,
  plot_text = FALSE,
  n_sim = 4,
  cols = c("#000000", "#cccccc", "ffffff", "ff0000", "0000ff"),
  chars = c(1, 1),
  sizes = c(0.1, 0.1),
  plot_pts = TRUE,
  ...
)
}
\arguments{
\item{input}{Use output of SRR simulation}

\item{p_thresh}{p threshold, default is 0.8}

\item{plot_text}{default is FALSE, in case resolution >> 10}

\item{n_sim}{TO ADD}

\item{cols}{colors, default colors = c("grey0", "grey80", "grey100", "red", "blue")}

\item{chars}{case, control}

\item{sizes}{case, control}

\item{plot_pts}{TO ADD, default is TRUE}

\item{...}{TO ADD}
}
\value{
None
}
\description{
Function to Estimate the Power of a Spatial Relative Risk using Simulated Data
}
\examples{
\donttest{
spatial_plots(input = sparrpowR_object, 
              p_thresh = 0.8, 
              #plot_text = T,
              plot_pts = T, 
              chars = c(4,5),
              sizes = c(0.5,0.5),
              cols = c("grey0", "grey80", "grey100", "red", "blue")
              )
}

}
