% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_power.R
\name{spatial_power}
\alias{spatial_power}
\title{spatial_power()}
\usage{
spatial_power(
  x_case,
  y_case,
  x_control = NULL,
  y_control = NULL,
  n_case = NULL,
  n_control = NULL,
  npc_control = NULL,
  r_case = NULL,
  r_control = NULL,
  s_case = NULL,
  s_control = NULL,
  l_case = NULL,
  l_control = NULL,
  e_control = NULL,
  sim_total = 1,
  samp_case = c("uniform", "MVN", "CSR", "IPP"),
  samp_control = c("uniform", "systematic", "MVN", "CSR", "IPP", "clustered"),
  upper_tail = 0.975,
  lower_tail = 0.025,
  win = spatstat::unit.square(),
  cascon = FALSE,
  resolution = 128,
  edge = "uniform",
  adapt = FALSE,
  h0 = NULL,
  verbose = TRUE,
  parallel = FALSE,
  n_core = NULL,
  ...
)
}
\arguments{
\item{x_case}{TO ADD}

\item{y_case}{TO ADD}

\item{x_control}{TO ADD}

\item{y_control}{TO ADD}

\item{n_case}{TO ADD}

\item{n_control}{TO ADD}

\item{npc_control}{TO ADD}

\item{r_case}{TO ADD}

\item{r_control}{TO ADD}

\item{s_case}{TO ADD}

\item{s_control}{TO ADD}

\item{l_case}{TO ADD}

\item{l_control}{TO ADD}

\item{e_control}{TO ADD}

\item{sim_total}{TO ADD}

\item{samp_case}{TO ADD}

\item{samp_control}{TO ADD}

\item{upper_tail}{TO ADD}

\item{lower_tail}{TO ADD}

\item{win}{TO ADD}

\item{cascon}{TO ADD}

\item{resolution}{TO ADD}

\item{edge}{TO ADD}

\item{adapt}{TO ADD}

\item{h0}{TO ADD}

\item{verbose}{TO ADD}

\item{parallel}{TO ADD}

\item{n_core}{TO ADD}

\item{...}{TO ADD}
}
\value{
Simulated point-level spatial data and local power.
}
\description{
Function to Estimate the Power of a Spatial Relative Risk using Simulated Data
}
\examples{
spatial_power(x_case = c(0.25),
              x_control = c(0.25),
              y_case = c(0.75),
              y_control = c(0.75),
              n_case = 10,
              n_control = 50,
              r_case = 0.1,
              r_control = 0.1,
              s_case = 0.05,
              s_control = 0.1,
              l_case = 200,
              l_control = 100,
              samp_case = "MVN", 
              samp_control = "MVN",
              npc_control = 100,
              e_control = 0,
              sim_total = 2,
              cascon = FALSE, 
              verbose = FALSE
              ) 

}
